Issues #23. Rename class.
