+++ modules/fraud/components/ValidateWithChangeStep.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace app\modules\fraud\components\validators\orders;
+namespace app\modules\fraud\components;
 
 use app\models\product\ProductSearch;
 use app\modules\fraud\components\FraudException;
@@ -10,10 +10,10 @@ use app\modules\fraud\models\FraudCheckStep;
 use app\modules\fraud\models\FraudCheckStepSearch;
 
 /**
- * Class ProductOrderValidatorWithChangeStep
- * @package app\modules\fraud\components\validators\orders
+ * Class ValidateWithChangeStep
+ * @package app\modules\fraud\components
  */
-class ValidateProductOrderWithChangeStep {
+class ValidateWithChangeStep {
 	private FraudValidator $validator;
 
 	public function __construct(FraudValidator $validator) {
@@ -21,15 +21,13 @@ class ValidateProductOrderWithChangeStep {
 	}
 
 	public function validate(int $entityId):void {
-		$existentOrder = (new ProductSearch())->getExistentSimcardOrder($entityId);
-		$validatorOrderRow = (new FraudCheckStepSearch())->getByOrderWithValidator($entityId, get_class($this->validator));
+		$validatorOrderRow = (new FraudCheckStepSearch())->getByEntityIdWithValidator($entityId, get_class($this->validator));
 		$this->validateWithCatch($validatorOrderRow);
 	}
 
 	public function repeatValidate(int $stepId):void {
 		$step = (new FraudCheckStepSearch())->getById($stepId);
 		$this->validateWithCatch($step);
-		$step->statusFail()->saveAndReturn();
 	}
 
 	protected function validateWithCatch(FraudCheckStep $step):void {
